fix: querydsl에서 변환할때 타입이 double, long이라 수정

+++ src/main/java/com/almondia/meca/category/controller/dto/CategoryWithHistoryResponseDto.java
@@ -23,6 +23,6 @@ public class CategoryWithHistoryResponseDto {
 	private final boolean isShared;
 	private final LocalDateTime createdAt;
 	private final LocalDateTime modifiedAt;
-	private final int scoreAvg;
-	private final int solveCount;
+	private final double scoreAvg;
+	private final long solveCount;
 }
